nvme: allow manually initializing the NVMe driver
We currently require the primary process to connect any
NVMe controller before secondary processes can connect
as well. This makes sense for PCIe NVMes, but not for
NVMe-oF. Users might want to use NVMe-oF initiator in
secondary processes with the primary process acting only
as a memory allocator.

spdk_nvme_driver_init() has been added to the public API.
It may be optionally called in the primary process to enable
secondary processes to connect NVMe controllers.

Within the same patch, we also call spdk_nvme_driver_init()
on bdev_nvme module init.

Fixes #768

Change-Id: Iea0075bbc6fe6668c5be578a8ffe3cdcce7cd008
Signed-off-by: Darek Stojaczyk <dariusz.stojaczyk@intel.com>