nvmf: complete all the I/Os when change the state of sgroup to PAUSED
For the nvme device, I/Os are completed asynchronously. So we need to check
the outstanding I/Os before putting IO channel when we hot remove the device.
We should be sure that all the I/Os have been completed when we change the
sgroup->state to PAUSED so that we can update the subsystem.

Fix #615 #755

Change-Id: I0f727a7bd0734fa9be1193e1f574892ab3e68b55
Signed-off-by: JinYu <jin.yu@intel.com>