test/rocksdb: add tracing
This will help collect evidence for the intermittent
RocksDB failures seen in the test pool.

The method for getting the name of the /dev/shm trace
file is pretty kludgy (grepping the log) but it's the
best I could come up with and is very simple.  Good
enough for now.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I7b3acbcb54033555ff42b3b21a09c798a87c481b