Update patch set 2
Patch Set 2:

(1 comment)

> Patch Set 2: Code-Review-1
> 
> (1 comment)
> 
> Hi Eugene,
> 
> I clearly see where you're going with this - and I like it.  I'd like to decouple the poll group traversal from the RPC stuff though.  We may have cases for other internal users of this data.  So it would be nice to have a generic "statistics" structure, and then user passes in this structure (or array of structures - one for each poll group) that the nvmf library fills in.  Then once it is done, RPC-specific code can convert it to JSON-RPC format.
> 
> Of course the thread names won't work in this case.  But I think that points to a deficiency elsewhere in SPDK.  Ben has been doing a lot of work in this area.  We may want to assign SPDK thread IDs, and then have an RPC like get_threads() which returns all of the thread IDs with their names.  This would allow you to cross-reference a thread ID from the nvmf traversal, with the actual name of the thread.
> 
> Number of poll group and number of transports are dynamic so that does make this more challenging.  Maybe we make the RPC specify a transport - and then the caller is responsible for calling the RPC once for each transport.  I know this pushes more logic to the RPC client - but in general we want to move as much of this type of complexity to the client where Python can easily manipulate and convert the data.
> 
> -Jim

Hi Jim,

Thanks for your comment.
In next patch sets I split this method in two. One is for generic NVMf subsystem statistics and another one is for transport-specific statistics.
I also decoupled gathering poll group statistics from writing JSON reply. For generic NVMf statistics it looks more or less fine.
But it is not that simple for transport-specific stats. I think it is impossible to create common statistics structure that fits all the transports. So, I had to create a separate structure for each transport and expose it in nvmf.h. As for me, this breaks incapsulation of transport layer. I'd be glad to hear any thoughts on this.

This exact change is not in the chain anymore after rework. I'll duplicate the answer in another change.

Patch-set: 2
Reviewer: Gerrit User 1015209 <1015209@d5d70762-12d0-45a1-890d-524b12d3f735>