Update patch set 3
Patch Set 3:

> Patch Set 2:
> 
> Sorry - meant to apply this to previous patch in this series.

Hi Jim,

Thanks for your comment.
The original change with your comment is not in the chain anymore after rework. So, I duplicate my answer here as well.

In next patch sets I split this method in two. One is for generic NVMf subsystem statistics and another is for transport-specific statistics.
I also decoupled gathering poll group statistics from writing JSON reply. For generic NVMf statistics it looks more or less fine.
But it is not that simple for transport-specific stats. I think it is impossible to create common statistics structure that fits all the transports. So, I had to create a separate structure for each transport and expose it in nvmf.h. As for me, this breaks incapsulation of transport layer. I'd be glad to hear any thought on this.

BR, Eugene.

Patch-set: 3
Reviewer: Gerrit User 1015209 <1015209@d5d70762-12d0-45a1-890d-524b12d3f735>