Update patch set 6
Patch Set 6:

The unfortunate part is that it does not catch the error with NVMe+NVMf that was fixed recently. Even though malloc and NVMe support iov type IOs, malloc does not fail in same way as NVMe did. This test passed on my setup when reverting the blobstore fix.

The simplification of the script is good change, but we will either need to use NVMe device here or think of better way to show previous issue.

Patch-set: 6
CC: Gerrit User 1011221 <1011221@d5d70762-12d0-45a1-890d-524b12d3f735>