iscsi: handle aborts for queued_datain_task that is completed
It's possible that we get a request to abort I/O just after
a queued_datain_task has completed, but before we've had a
chance to remove it from the queued_datain_tasks TAILQ.
_iscsi_conn_abort_queued_datain_task wasn't accounting for
that which would result in an infinite loop.

Fixes issue #725.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Tested-by: Jane Lusby
Change-Id: Id8b3eb029e3d097e52a0af144ec8735034a97050
Reviewed on: https://review.gerrithub.io/c/spdk/spdk/+/450301 (master)