Update patch set 2
Patch Set 2:

Ok so I've read this a bunch of times. SPDK doesn't use signed integers much, so no-strict-overflow/wrapv are unlikely to cause any performance impacts at all. That said, we certainly don't need both no-strict-overflow and wrapv - those are two different approaches for solving the same problem and we only need to choose one. There is a third option, too, which is Wstrict-overflow=5. That simply won't let any code that could induce a signed integer overflow  even compile. That may actually be the best solution for us, given that signed integers in SPDK are rare.

The no-delete-null-pointer option seems pointless in our environment, but if it doesn't hurt performance then I don't see it being that bad.

Patch-set: 2