Update patch set 15
Patch Set 15:

(25 comments)

Hi Anil,

In addition to the comments for the patch, I have a preliminary question.

For RDMA and TCP, NVMe-oF target runs the acceptor poller on a single core.
When NVMe-oF target receives incoming connection, it allocates a single core to the connection by round robin manner.
All qpairs for the connection will be processed in the core.

But for FC, acceptor of the NVMe-oF target do nothing, and I haven't found thread management in CASS and CIOC.

Will you explain about the thread management for FC association and connection?

Thanks,
Shuhei

Patch-set: 15
Reviewer: Gerrit User 1012251 <1012251@d5d70762-12d0-45a1-890d-524b12d3f735>