bdev: adjust the maximum number of supported children IOV based on data buffer
Here is the an example to describe existing issue:

There is a Write request with 64KiB data length, and this IO is cross the IO
boundary.  We assume that the parent IO will have 2 children requests, one is
33KiB length, the other one is 31KiB.  Here is the view of parent iovs, the
first 33KiB length data has 33 iovs:

iov.[0].iov_length = 1024;
.
.
iov.[31].iov_length = 256;
iov.[32].iov_length = 768;
.
.
iov.[64].iov_length = 1024;

In function _spdk_bdev_io_split(), then you can see that for the 33KiB length
child request, exiting code will return error because the last payload is not
512 aligned, however, if we increase the BDEV_IO_NUM_CHILD_IOV to a bigger value,
such as 128, it will pass the split progress without error, and then submit the
request to the backend, if the backend device is NVMe with hardware SGL support,
this is totally fine.  But what if the backend is AIO, so for this case the 33KiB
child request will go through spdk_bdev_io_get_buf() again and copy all the unaligned
iovs into a huge data buffer.

Existing framework can address the issue if we increase BDEV_IO_NUM_CHILD_IOV
to a bigger value.

Change-Id: I96ebdf29829d86f9b38fab28a7406eedc9fa44ef
Signed-off-by: Changpeng Liu <changpeng.liu@intel.com>