Update patch set 5
Patch Set 5: Code-Review+2

This looks good for now, but long term consider a slightly different approach. Create a single object (can be global) that represents the vhost target. Register that as an io_device and have the channels for that io_device be the poll groups.

The advantage of this approach is when you go to shut down a poll group (or perform some global operation), you may need perform operations on the thread that owns the poll group, and there is easy infrastructure to do all of the message passing via spdk_for_each_channel(). This is how NVMe-oF works. You also wouldn't need to keep your own global list of poll groups anymore.

Patch-set: 5
Reviewer: Gerrit User 1010531 <1010531@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=+2