nvme: spdk_nvme_ctrlr_alloc_io_qpair extensions
Adds fields to spdk_nvme_io_qpair_opts to specify the memory location of
an I/O queue pair using storage addresses placed within this structure.

Four new fields allow one to specify the starting address
of the submission and completion queue respectively.
If a vaddr value is NULL, SPDK allocates the memory used.
If a paddr value is specified, SPDK uses the vaddr and paddr passed
but does NOT assume said memory is registered.  These can be used
for device testing, or to satisfy potential performance issues.

SPDK assumes the memory passed is both virtually and physically contiguous.

If these fields are used, SPDK will NOT impose any restriction
on the number of elements in the queues.

These fields are only used by PCIe attached NVMe devices.  They
are presently ignored for other transports.

This function can only be used with PCIe controllers.  Other transports return NULL.

This function is primarily intended for device testing.

Signed-off-by: James Bergsten <jamesx.bergsten@intel.com>
Change-Id: Ibfab3939eefe48109335f43a1167082dd4865e7c