vhost: register vvu PCI device as a DMA-calable device
In case of vhost-user with vvu transport, the master vhost memory
regions are exposed to the slave guest as regions of the memory space of
the virtio-vhost-user device. In case of vfio no-IOMMU mode, where SPDK
stores the VA-to-PA translations in the vtophys map, the physical
addresses associated with the vhost memory regions point out to the
virtio-vhost-user device memory regions. Therefore, it is necessary that
the virtio-vhost-user device is registered to the SPDK's internal PCI
device list called `g_vtophys_pci_devices` so that SPDK can find the
VA-to-PA translations.

The virtio-vhost-user device is being handled by the virtio-vhost-user
driver in DPDK.  In order to find the `rte_pci_device` instance
associated with the virtio-vhost-user device, we use DPDK's
`rte_pci_dev_iterate()` function. This function searches in DPDK's
internal PCI device list using a PCI BDF address and, in case of a
match, it returns the corresponding `rte_device` instance. Also, DPDK's
`rte_bus_pci.h` header file has been included because it contains the
definitions of `struct rte_pci_device` and `RTE_DEV_TO_PCI()`.

Change-Id: Iad2522bee97e9e3ff9cb2e1782e99d31b32f6620
Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>