vhost: restrict the 2MB mmap_size alignment check
In case of the vvu transport, the vhost memory regions exposed via the
memory space of the virtio-vhost-user PCI device do not necessarily have
2MB aligned mmap_size, because they are not necessarily hugepage-backed.
This patch removes the check about the 2MB mmap_size alignment of the
vhost memory regions in case of the vvu transport. We keep the check for
the case of the AF_UNIX transport, though it should be completely
removed at some point along with the demand for hugepage-backend guest
memory.

The 2MB mmap_size alignment check is just a validation point that the
vhost memory regions are hugepage-backed. However, hugepage-backed
memory is more of an efficiency enhancement rather than a prerequisite
for the vhost-user protocol. The SPDK vhost target would still work if
the master guest memory was backed by normal 4KB pages, provided that we
previously had enabled registering non 2MB-aligned memory regions in the
SPDK memory maps (it will be done at some point).

Change-Id: I2bbc6b65aa53d8af8b5a4f4665965bb95eae5c91
Signed-off-by: Nikos Dragazis <ndragazis@arrikto.com>