nvme: add nvme process_completions variants with tsc
This will allow the driver to make decisions such as
cq doorbell coalescing based on the current tsc.

Note that we have the user pass in the tsc, rather
than having the driver explicitly get it, since getting
the tsc can be expensive, and the calling application
may already have a fairly recent tsc collected for other
reasons.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I063b38011a487df41ca6cb38c06d16b025032803