Update patch set 4
Patch Set 4:

So the issue in #625 is for example in iSCSI target, iSCSI target requests 64KB I/O but if the buffer is unaligned, the size of the buffer becomes more than 64KB and cannot be allocated?

I haven't seen your another patch yet but #1 may be the one we should take.

Patch-set: 4