blobfs: Only set the is_deleted attribute once
Issue: If we call spdk_fs_delete_file several times when
the file is still open (with ref_count > 0). Then the second
time, the "is_deleted" flag will be set again, and it is
unnecessary, since it is already set by the first time.

And this patch can be used to solve this issue.

Signed-off-by: Ziye Yang <ziye.yang@intel.com>
Change-Id: I0d996fedc4f5f50d04e433a873fb8bd3e832939c