Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> (1 comment)
> 
> Awesome Seth, just one small thing.  Also, in the commit msg you mention preferably NULL bdev, there's not restrictions or anything in here for that.  Why mention it? Just to avoid other latency entagelemts of various bdev modules?

Yeah, I figured that if you are trying to simulate the latency of an NVMe drive, putting a delay bdev on top of a Null bdev will give you the most accurate representation of that drive. Not saying that this model has a super high degree of fidelity to an actual drive right not, but I think putting it on top of any other bdev could increase the disparity.

Patch-set: 4
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>