Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review-1
> 
> Hi Seth,
> 
> Why do you use average latency? Unfortunately average latency is not intuitive for me yet. I think the following may be easier to understand?
> 
> p99 latency: the worst latency observed by 99% of requests.
> worst latency: the worst latency observed by the top 1% of requests.
> 
> p99 latency should be the maximum value if we ignore the top 1%?
> 
> And should we check p99 lantecy is (much) smaller than the worst latency?
> 
> Or will you add some more explanation?

Oh I see how the way I am using p99 could be confusing. Typically p99 means that 99% of I/O fall within that latency range. What I am using it to mean is that is the latency number that 1% of I/O have to be slower than that latency. So I submit 1% of I/O with a delay = to p99 latency ensuring that their latency after being polled will be greater than that number.
When I say average latency, I mean if you were to run FIO or another tool on a drive, it would report back that number as the mean latency. I submit most I/O with a delay = to average latency so that you hopefully see that number coming through as the average latency of I/O when running a tool like fio to profile a NULL bdev. 
It's sort of my way of trying to emulate two major drive characteristics without the added complexity creating a matrix of lists to represent a normalized distribution of I/O latencies.
I will update some of the help output and docs to make my intentions more clear.

Patch-set: 4
Reviewer: Gerrit User 1011223 <1011223@d5d70762-12d0-45a1-890d-524b12d3f735>