examples/nvme: add overlap test
This is an example application that shows the cost of
completing two read I/O to the same LBA, when delaying
CQ doorbell writes for the first completion for 1
second.  Some versions of FW on some SSD models seem
to not start executing the second I/O until the CQ
doorbell has been written to acknowledge the first I/O
completion.

Signed-off-by: Jim Harris <james.r.harris@intel.com>
Change-Id: I7a0f295d9efaf34683f46c761c4d69eecdc8a4e5