dif: Use struct dif_sgl to simplify to create iovec array
struct dif_sgl had been used only to iterate the passed iovec
array. This patch utilizes struct dif_sgl and adds a helper
function _dif_sgl_append() to simplify to create iovec array.

Besides, iovcnt has been used to mean the number of elements
in the iovec array widely in the DIF library. So this patch
changes the name of it from num_iovs to iovcnt.

Change-Id: Ib9c5a56bb7839a0415c2d1647a26b24ef3eb0455
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>