dif: Use _dif_sgl_append in spdk_dif_set_md_interleave_iovs
This patch uses _dif_sgl_append in spdk_dif_set_md_interleave_iovs
to simplify to create iovec array.

Besides, iovcnt has been used to mean the number of elements
in the iovec array widely in the DIF library. So this patch
changes the name of it from num_iovs to iovcnt.

Change-Id: Ib9c5a56bb7839a0415c2d1647a26b24ef3eb0455
Signed-off-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>