test/bdevio: app now can wait for RPC before tests
Added '-w' argument to the bdevio app, so that instead
of performing the tests immidietly - it awaits RPC.

Patches further in the series will add new RPC specific
to this application to actually run the tests.

After a lot of changes to blockdev.sh, we might want to
remove this argument and just use it as default.

Change-Id: Id5ca24c4da2a782cd204af354df966c478af1803
Signed-off-by: Tomasz Zawadzki <tomasz.zawadzki@intel.com>