Update patch set 1
Patch Set 1:

Are these scripts still needed if you can run each vhost test in isolation?

If there are some helper scripts that would still make all of this easier, I'd be fine with keeping some version of these scripts in the repository.  But I'm guessing they would need to be refactored based on a lot of the other changes Ben's making later in this series.

I vote for merging this patch as-is, and Karol push a patch later that contains updated versions of these scripts (possibly in a different directory).

Patch-set: 1
CC: Gerrit User 1011222 <1011222@d5d70762-12d0-45a1-890d-524b12d3f735>