test: run bdev_raid.sh nightly instead of per patch
Existing raid UT already checks the logic in the raid
unmap code to intelligently send IOs to member disks.

The additional value of bdev_raid.sh is in doing a real
data validation on writes as well as confirmation that
unmaps really result in 0's following the action. Both of
these seem worth keeping however with the UT in place to
check the logic they are a better fit for nightly testing.

Change-Id: I1a758f4681d77dbddd4185162b0dd1f46525cabd
Signed-off-by: paul luse <paul.e.luse@intel.com>
Reviewed-on: https://review.gerrithub.io/c/spdk/spdk/+/454668
Tested-by: SPDK CI Jenkins <sys_sgci@intel.com>
Reviewed-by: Jim Harris <james.r.harris@intel.com>
Reviewed-by: Shuhei Matsumoto <shuhei.matsumoto.xt@hitachi.com>