bdev/compress: updates to the poller function
Two quick updates, support dequeueing more than 1 IO at a time
which is now more likely given that we can queue up compression
operations (others can come in qhile one is on the queue). Dropped
the max inflight value as it was in there as a palceholder and now
it makes sense to give it a sane value.

Also updated the value passed to reduce to signal an error.

UT for the poller will come in the next patch...

Change-Id: Ic7f7fcaac2182153e07bc1a64bddf76a33c5751d
Signed-off-by: paul luse <paul.e.luse@intel.com>