Update patch set 8
Patch Set 8: Code-Review-2

The iSCSI bdev on our test machines can barely reach the hardcoded QoS limit and we can't fix it by increasing the allowed QoS deviation. We can still end up not running the test if the device happens to be to slow, which in my opinion is much worse than not running it at all. That's because instead of knowing you don't run any QoS tests, you might think you run them, but in reality you don't - because the test is being skipped all the time.

We should always avoid scenarios where we end up skipping the test. In this case it would be fairly simple - instead of setting an arbitrary QoS limit, we could set a limit that's a % of the actual disk performance. Then we don't have to check if the disk is fast enough and the test is always run. Please see this patch:

https://review.gerrithub.io/c/spdk/spdk/+/451887 test/qos: set qos limits to a % of the maximum disk performance

Patch-set: 8
Reviewer: Gerrit User 1011218 <1011218@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-2