Update patch set 5
Patch Set 5: Code-Review-1

Hi Ziye, not sure if you saw my earlier comment as I didn't vote but it was the same as Ben's. We talked about how to strike a balance between testing all of the combinations of targets and bdev modules in the test team that standardized on autotest and moved us to Jenkins over the last few years. It was decided at that time that we'd only test "end to end" with the most common building blocks and we'd rely on more specific testing to assure individual modules are validated. If we tested every combination of everything our per patch testing would take too long. I think this is a better general topic for a community meeting discussion as opposed to a patch but for this one specifically, cryotp is already tested with bdevperf, fio and bdevio which includes data integrity testing. We can probably "beef up" bdev testing even more and get that worked leveraged over all bdev modules with minimal overhead.

Patch-set: 5
Reviewer: Gerrit User 1011289 <1011289@d5d70762-12d0-45a1-890d-524b12d3f735>
Label: Code-Review=-1